requirements.txt
paho-mqtt
python-decouple
RPi.GPIO